requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
jinja2>=3.1.0
requests>=2.28.0
tldextract>=3.4.0
python-multipart>=0.0.6
aiohttp>=3.8.0
python-whois>=0.8.0
beautifulsoup4>=4.12.0
